+++ R/dotDensityLayer.R
@@ -13,7 +13,6 @@
 #' @param var name of the numeric variable to plot.
 #' @param n one dot on the map represents n (in var units).
 #' @param col color of the points.
-#' @param iter defunct.
 #' @param pch symbol to use: \link{points}.
 #' @param cex size of the symbols
 #' @param type points allocation method: "random" or "regular" (see Details).
@@ -50,12 +49,9 @@ dotDensityLayer <- function(x, spdf, df, spdfid = NULL, dfid = NULL, var,
                             legend.cex = 0.6,
                             legend.col = "black",
                             legend.frame = TRUE,
-                            add = TRUE, iter){
+                            add = TRUE){
   
-  if (!missing(iter)){
-      stop("iter argument is defunct", call. = FALSE)
-  }
-  
+
   if (missing(x)){
     x <- convertToSf(spdf = spdf, df = df, spdfid = spdfid, dfid = dfid)
   }