+++ R/getTiles.R
@@ -3,7 +3,6 @@
 #' @description Get map tiles based on a spatial object extent. Maps can be 
 #' fetched from various open map servers.
 #' @param x an sf object, a simple feature collection or a Spatial*DataFrame.
-#' @param spdf  defunct.
 #' @param type the tile server from which to get the map. See Details for providers.
 #' For other sources use a list: type = list(src = "name of the source" , 
 #' q = "tiles address", sub = "subdomains", cit = "how to cite the tiles"). See Examples.
@@ -82,16 +81,11 @@
 #' }
 getTiles <- function(x, type = "OpenStreetMap", zoom = NULL, crop = FALSE, 
                      verbose = FALSE, apikey = NA, cachedir = FALSE, 
-                     forceDownload = FALSE, spdf){
+                     forceDownload = FALSE){
   .Deprecated(
     new = "get_tiles", package = "maptiles", 
     msg = "'getTiles' is deprecated.\nUse 'maptiles::get_tiles()' instead."
   )
-  # defunct
-  if(!missing(spdf)){
-    stop("spdf is defunct; use x instead.", call. = FALSE)
-    x <- spdf
-  }
   # test for sp
   if(methods::is(x,"Spatial") == TRUE){
     x <- convertToSf(spdf = x)